requirements.txt
pandas
numpy
PyPDF2
pikepdf
dash
pdfplumber
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
wordsegment
nltk
pyspellchecker
sqlalchemy
tqdm
python-dotenv
PyYAML
PyPDF2
pypdf
streamlit
gspread
gspread_dataframe